+++ packages/amplify-codegen/src/commands/models.js
@@ -29,6 +29,19 @@ const readFeatureFlag = key => {
   return flagValue;
 };
 
+/**
+ * Returns feature flag value, default to `1`
+ * @param {string} key feature flag id
+ * @returns
+ */
+const readNumericFeatureFlag = key => {
+  try {
+    return FeatureFlags.getNumber(key);
+  } catch (err) {
+    return 1;
+  }
+};
+
 async function generateModels(context) {
   // steps:
   // 1. Load the schema and validate using transformer
@@ -72,8 +85,8 @@ async function generateModels(context) {
 
   const generateIndexRules = readFeatureFlag('codegen.generateIndexRules');
   const emitAuthProvider = readFeatureFlag('codegen.emitAuthProvider');
-  const usePipelinedTransformer = readFeatureFlag('graphQLTransformer.useExperimentalPipelinedTransformer');
-  const improvePluralization = readFeatureFlag('graphQLTransformer.improvePluralization');
+  const usePipelinedTransformer = readFeatureFlag('graphQLTransformer.useExperimentalPipelinedTransformer')
+  const transformerVersion = readNumericFeatureFlag('graphQLTransformer.transformerVersion');
 
   let enableDartNullSafety = readFeatureFlag('codegen.enableDartNullSafety');
 
@@ -105,7 +118,7 @@ async function generateModels(context) {
       enableDartNullSafety,
       handleListNullabilityTransparently,
       usePipelinedTransformer,
-      improvePluralization
+      transformerVersion,
     },
   });