+++ packages/amplify-codegen/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.26.10](https://github.com/aws-amplify/amplify-codegen/compare/amplify-codegen@2.26.9...amplify-codegen@2.26.10) (2021-11-03)
+
+
+### Reverts
+
+* Revert "Feature: Use improved pluralization" ([a891a4f](https://github.com/aws-amplify/amplify-codegen/commit/a891a4f5820fcdb9610261b01ecc4d1211cf4e9f))
+
+
+
+
+
+## [2.26.9](https://github.com/aws-amplify/amplify-codegen/compare/amplify-codegen@2.26.8...amplify-codegen@2.26.9) (2021-10-27)
+
+**Note:** Version bump only for package amplify-codegen
+
+
+
+
+
+## [2.26.8](https://github.com/aws-amplify/amplify-codegen/compare/amplify-codegen@2.26.7...amplify-codegen@2.26.8) (2021-10-21)
+
+**Note:** Version bump only for package amplify-codegen
+
+
+
+
+
+## [2.26.7](https://github.com/aws-amplify/amplify-codegen/compare/amplify-codegen@2.26.6...amplify-codegen@2.26.7) (2021-10-05)
+
+**Note:** Version bump only for package amplify-codegen
+
+
+
+
+
 ## [2.26.6](https://github.com/aws-amplify/amplify-codegen/compare/amplify-codegen@2.26.5...amplify-codegen@2.26.6) (2021-09-29)
 
 **Note:** Version bump only for package amplify-codegen